// guipmc-casuistica/src/app/casuistica/models/control-figures/form-control-figure.ts
export class FormControlFigure {
    fromMonth: string;
    fromYear: string;
    toMonth: string;
    toYear: string;
    cveTipoArchivo: string;
    delRegPat: boolean;
    page: number;
    isPdfReport: boolean;
    desArchivo: string;
    cveDelegation: string;
    cveSubdelegation: string;
    desSubdelegation: string;
    desDelegation: string;
    cveCasoRegistro: string;
  tipoReporte: string;
  isDelegacional: boolean;
}